+++ public/lib/urls.js
@@ -25,7 +25,7 @@ const loopUrlData = (urlList) => {
 const appendUrlATags = (urlInfo) => {
   const { id, long_url, visits } = urlInfo;
   const urlATag = $(`
-    <a href="/${id}" target="_blank">/${id} => (${long_url})</a>
+    <a href="${id}" target="_blank">/${id} => (${long_url})</a>
     <p> visits: ${visits} </p> </br>
   `)
   $('#urls').append(urlATag)
